pyproject.toml
[project]
name = "rpisonyremote"
version = "0.1.0"
description = "RPi Sony Remote"
authors = [
    {name = "[SHS]MAD", email = "shsmad@gmail.com"},
]
dependencies = [
    "bleak",
    "luma.oled",
    "RPI.GPIO",
    "Pillow",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

# [project.urls]
# repository = "https://coderepo.corp.tander.ru/remeslennikov_ai/jiraplus"

# [[tool.pdm.source]]
# name = "pypi"
# url = "https://repo.corp.tander.ru/repository/webdevrepo-pypi-all/simple"
# verify_ssl = false

[tool.pdm.dev-dependencies]
linter = [
    "ruff>=0.1.6",
    "mypy>=1.7.0",
]

dev = [
    "wheel>=0.41.3",
    "ipdb>=0.13.13",
    "ipython>=8.17.2",
    #"pyopenssl",
]

[tool.ruff]
target-version = "py39"
exclude = [
    ".tox",
    ".git",
    "__pycache__",
    "docs/source/conf.py",
    "old",
    "build",
    "dist",
    "*_pb2*.py",
]
ignore = [
    "A003",
    "D107",
    "D202",
    "D203",
    "D401",
    # for ruff
    "COM812",
]
line-length = 120
select = [
    "A", # builtins
    "B", # bugbear
    "C4", # comprehensions
    "COM", # commas
    # "D",   # pydocstyle
    "E", # pycodestyle
    "F", # pyflakes
    "I", # isort
    "PT", # pytest-style
    "S", # bandit
    "SIM", # simplify
    "T1", # debugger
    "T2", # print
    "UP", # pyupgrade
    "W", # pycodestyle
]

# Avoid trying to fix flake8-bugbear (`B`) violations.
unfixable = ["B"]

[tool.ruff.per-file-ignores]
"test/*.py" = ["S101"]

[tool.ruff.isort]
lines-between-types = 1
known-first-party = ["test"]
known-local-folder = ["src"]
force-single-line = false


[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
# disallow_subclassing_any = true
disallow_incomplete_defs = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
no_implicit_reexport = true
# allow_redefinition = true
pretty = true
follow_imports = "skip" # skip/silent
ignore_missing_imports = true
show_error_codes = true
strict_optional = true
warn_no_return = true
# for strict mypy: (this is the tricky one :-))
disallow_untyped_defs = true


[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true


[tool.pdm.scripts]
lint = "ruff check app"
